pyproject.toml
[tool.poetry]
name = "foreshadwownba-dashboard-monitoring"
version = "0.1.0"
description = ""
authors = ["Yannick Flores <yannick.flores1992@gmail.com>"]
readme = "README.md"
packages = [{include = "foreshadwownba_dashboard_monitoring"}]

[tool.poetry.dependencies]
python = "^3.11"
ipykernel = "6.29.5"
pandas = "2.2.3"
shiny = "1.2.1"
htmltools = "0.6.0"
python-dotenv  = "1.0.1"
faicons = "*"
seaborn = "*"
scikit-learn = "1.6.0"

[tool.poetry.dev-dependencies]
coverage = {extras = ["toml"], version = "6.5.*"}
black = {version = "22.10.*", allow-prereleases = true}
pytest = "7.2.*"
pytest-html = "3.1.*"
pytest-cov = "2.12.*"
pytest-playwright = "*"
py = "1.11.*"
pylint = "2.17.*"
python-decouple = "3.6"
mkdocs = "1.4.*"
importlib-resources="6.0.1"
pandantic = "0.3.1"
skimpy = "0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
testpaths = "tests"
addopts = [
    "--import-mode=importlib",
]
